Function to build/style cards
Removed the .replace on line 89 as it was causing a critical error for me... not totally sure why or what it does. Moved the card build function to the end to minimize conflicts hopefully. Added local storage of result data on 127, card build call on 128. Also styles in css. And some additional containers in html for flex-boxing.